op-proposer: Reorganize packge
